pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ConvoCart"
version = "0.1.0"
description = "ConvoCart with AI Agent Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "transitions>=0.9.3",
    "uvicorn>=0.34.2",
    "phonenumbers>=8.13.37",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "google-generativeai>=0.8.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "typing-extensions>=4.13.2",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "twilio>=9.0.0",
    "starlette-csrf>=3.0.0",
    "fastapi-csrf-protect>=1.0.7",
    "diskcache>=5.6.3",
    "celery>=5.5.3",
    "redis>=5.0.0",
    "setuptools>=80.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "node_modules*", ".venv*"]
